Add repository and observable repository
This change incorporates some additional common functionality. Aside
from the original mongo-backed queue, this now includes a simple
repository implementation along with an observable repository that
takes advantage of MongoDB's change streams.

The intent is to provide more common functionality within this project,
so it can be consumed as a submodule in multiple other projects that
depend on it.